Move event.py to the player package.
